Refactor QuestionnaireActivity and QuestionnaireViewModel classes
