+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/main/AppMainViewModel.kt
@@ -77,7 +77,7 @@ import org.smartregister.fhircore.engine.util.extension.showToast
 import org.smartregister.fhircore.engine.util.extension.tryParse
 import org.smartregister.fhircore.quest.navigation.MainNavigationScreen
 import org.smartregister.fhircore.quest.navigation.NavigationArg
-import org.smartregister.fhircore.quest.ui.questionnaire.QuestionnaireNewActivity
+import org.smartregister.fhircore.quest.ui.questionnaire.QuestionnaireActivity
 import org.smartregister.fhircore.quest.ui.report.measure.worker.MeasureReportMonthPeriodWorker
 import org.smartregister.fhircore.quest.ui.shared.QuestionnaireHandler
 import org.smartregister.fhircore.quest.ui.shared.models.QuestionnaireSubmission
@@ -242,7 +242,7 @@ constructor(
           extraIntentBundle =
             bundleOf(
               Pair(
-                QuestionnaireNewActivity.QUESTIONNAIRE_POPULATION_RESOURCES,
+                QuestionnaireActivity.QUESTIONNAIRE_POPULATION_RESOURCES,
                 arrayListOf(location),
               ),
             ),
